chore(tracker): int32 dispatch counters, isolated RPC test data, CHANGELOG (#2022)
TickResult counters are int32 at the source (they are the response's
type), removing the int->int32 narrowing gosec flagged. The RPC tick
test deletes its connection first so no dispatch row survives a prior
run. CHANGELOG Unreleased entry.

+++ internal/server/tracker_dispatch_server.go
@@ -61,9 +61,9 @@ func (s *ContainerServer) DispatchTrackerIssues(ctx context.Context, req *pb.Dis
 	}
 
 	out := &pb.DispatchTrackerIssuesResponse{
-		SkippedNeedsApproval: int32(res.SkippedNeedsApproval),
-		SkippedActive:        int32(res.SkippedActive),
-		SkippedUnrouted:      int32(res.SkippedUnrouted),
+		SkippedNeedsApproval: res.SkippedNeedsApproval,
+		SkippedActive:        res.SkippedActive,
+		SkippedUnrouted:      res.SkippedUnrouted,
 	}
 	for i := range res.Started {
 		out.Started = append(out.Started, toProtoTrackerDispatch(&res.Started[i]))
